+++ src/epaper/__main__.py
@@ -10,7 +10,7 @@ os.environ.setdefault("GPIOZERO_PIN_FACTORY", config().get("gpiozero", {}).get("
 
 from epaper.display import Display
 from epaper.price.client import BitcoinPriceClient
-from epaper.price.extractor import PriceExtractor
+from epaper.price.price_extractor import PriceExtractor
 from epaper.price_ticker import PriceTicker
 from epaper.utils.graceful_shutdown import GracefulShutdown
 from epaper.utils.watchdog import sd_notify
